requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
pandas>=1.2.0
psycopg2-binary>=2.9.3
python-dotenv>=0.19.0 